+++ packages/kiwi-core/src/Button/tests/button.test.js
@@ -1,5 +1,5 @@
 import Button from '@/packages/kiwi-core/src/Button'
-import { render, defaultProviders } from '@/packages/kiwi-core/src/test-utils'
+import { render, defaultProviders } from '@/tests/test-utils'
 
 const renderComponent = (props) => {
   const base = {
