+++ packages/kiwi-core/src/CloseButton/tests/CloseButton.test.js
@@ -1,5 +1,5 @@
 import CloseButton from '../'
-import { render } from '@/packages/kiwi-core/src/test-utils'
+import { render } from '@/tests/test-utils'
 
 const renderComponent = (props) => {
   const base = {
