+++ packages/kiwi-core/src/Divider/tests/Divider.test.js
@@ -1,5 +1,5 @@
 import Divider from '../'
-import { render } from '@/packages/kiwi-core/src/test-utils'
+import { render } from '@/tests/test-utils'
 
 const renderComponent = (props) => {
   const inlineAttrs = (props && props.inlineAttrs) || ''
