+++ packages/kiwi-core/src/Flex/tests/Flex.test.js
@@ -1,5 +1,5 @@
 import Flex from '../'
-import { render } from '@/packages/kiwi-core/src/test-utils'
+import { render } from '@/tests/test-utils'
 
 const renderComponent = (props) => {
   const inlineAttrs = (props && props.inlineAttrs) || ''
