+++ packages/kiwi-core/src/Fragment/tests/Fragment.test.js
@@ -1,5 +1,5 @@
 import Fragment from '../'
-import { render } from '@/packages/kiwi-core/src/test-utils'
+import { render } from '@/tests/test-utils'
 
 const renderComponent = (props) => {
   const base = {
