+++ packages/kiwi-core/src/Heading/tests/Heading.test.js
@@ -1,5 +1,5 @@
 import Heading from '../'
-import { render } from '@/packages/kiwi-core/src/test-utils'
+import { render } from '@/tests/test-utils'
 
 const renderComponent = (props) => {
   const base = {
